requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.1
aiohttp-cors==0.7.0
aiokafka==0.12.0
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==5.0.1
attrs==24.2.0
avro==1.12.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorlog==6.9.0
confluent-kafka==2.6.0
croniter==2.0.7
dnspython==2.7.0
email_validator==2.2.0
fastavro==1.9.7
faust-streaming==0.11.3
fhir.resources==7.1.0
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
intervaltree==3.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mode-streaming==0.4.1
multidict==6.1.0
mypy-extensions==1.0.0
opentracing==2.4.0
packaging==24.2
propcache==0.2.0
protobuf==5.28.3
psycopg2==2.9.10
pydantic==2.10.1
pydantic_core==2.27.1
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
terminaltables==3.1.10
typing_extensions==4.12.2
urllib3==2.2.3
venusian==3.1.0
yarl==1.17.1
